requirements.txt
annotated-types==0.6.0
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
docstring_parser==0.16
Flask==3.0.3
Flask-SocketIO==5.3.6
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
rsa==4.9
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
Flask
flask-socketio
google-cloud-storage
google-cloud-vision
google-cloud-speech
google-cloud-texttospeech
google-cloud-language
opencv-python-headless
tensorflow
python-dotenv
google-generativeai
